backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
buildozer==1.5.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distlib==0.4.0
docutils==0.22.3
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
ptyprocess==0.7.0
Pygments==2.19.2
PyJWT==2.12.1
pyodbc==5.3.0
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzbar==0.1.9
qrcode==8.2
requests==2.32.5
sh==2.2.2
six==1.17.0
tzdata==2025.2
urllib3==2.6.1
virtualenv==20.35.4
Werkzeug==3.1.3
